updated deleteProduct controller

+++ backend/controllers/product.controller.js
@@ -51,12 +51,16 @@ export const updateProduct = async (req, res) => {
 
 export const deleteProduct = async (req, res) => {
     const { id } = req.params
+
+    if (!mongoose.Types.ObjectId.isValid(id)) {
+        return res.status(400).json({ success: false, message: "Invalid id" });
+    }
     try {
         await Product.findByIdAndDelete(id);
         console.log("Product deleted successfully");
         res.status(200).json({ success: true, msg: "Product deleted successfully" })
     }
     catch (err) {
-        res.status(404).json({ success: false, msg: "Product not found" });
+        res.status(500).json({ success: false, msg: "Server Error" });
     }
 }
